udemy_bots/udemy_final/requirements.txt
asyncpg
aiogram>=2.11.2,<3.0
python-dotenv
